+++ lib/world_nailao_flutter_utils.dart
@@ -3,7 +3,10 @@ library utils;
 export '/src/dio_utils/dio_client.dart' show DioClient;
 export '/src/dio_utils/interceptor/check_response_state_interceptor.dart'
     show CheckResponseStateInterceptor;
-export '/src/dio_utils/interceptor/log_print_interceptor.dart' show LibLogInterceptor;
+export '/src/dio_utils/interceptor/log_print_interceptor.dart'
+    show LibLogInterceptor;
+
+export '/src/dio_utils/exception/api_exception.dart' show ApiException;
 
 export '/src/dio_utils/dio_utils.dart' show DioUtils;
 export '/src/is_empty/is_empty.dart' show isEmpty;